requirements-api.txt
# ==============================================
# Healthcare No-Show API - Dependencies
# ==============================================

# Core API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ML Dependencies
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0

# Optional: XGBoost (if using XGBoost model)
xgboost>=2.0.0

# HTTP Client (for testing)
httpx>=0.25.0
requests>=2.31.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Logging & Monitoring
python-json-logger>=2.0.0
colorlog>=6.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
isort>=5.12.0

# Production Server
gunicorn>=21.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0